+++ src/main/java/shop/yesaladin/shop/product/domain/model/search/SearchedProductPublisher.java
@@ -1,0 +1,16 @@
+package shop.yesaladin.shop.product.domain.model.search;
+
+import lombok.AllArgsConstructor;
+import lombok.Getter;
+import org.springframework.data.elasticsearch.annotations.Field;
+import org.springframework.data.elasticsearch.annotations.FieldType;
+
+@Getter
+@AllArgsConstructor
+public class SearchedProductPublisher {
+
+        @Field(name = "id", type = FieldType.Integer)
+        private Long id;
+        @Field(name = "name", type = FieldType.Keyword)
+        private String name;
+}
