+++ src/test/java/shop/yesaladin/shop/product/service/impl/SearchProductServiceImplTest.java
@@ -1,8 +1,8 @@
 package shop.yesaladin.shop.product.service.impl;
 
+import static org.assertj.core.api.Assertions.assertThat;
 import static org.mockito.Mockito.atLeastOnce;
 import static org.mockito.Mockito.verify;
-import static org.assertj.core.api.Assertions.assertThat;
 
 import java.time.LocalDate;
 import java.util.List;
@@ -10,10 +10,10 @@ import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.mockito.Mockito;
-import shop.yesaladin.shop.product.domain.model.SearchedProduct.SearchedAuthor;
-import shop.yesaladin.shop.product.domain.model.SearchedProduct.SearchedCategories;
-import shop.yesaladin.shop.product.domain.model.SearchedProduct.SearchedFile;
-import shop.yesaladin.shop.product.domain.model.SearchedProduct.SearchedTags;
+import shop.yesaladin.shop.product.domain.model.search.SearchedProductAuthor;
+import shop.yesaladin.shop.product.domain.model.search.SearchedProductCategory;
+import shop.yesaladin.shop.product.domain.model.search.SearchedProductFile;
+import shop.yesaladin.shop.product.domain.model.search.SearchedProductTag;
 import shop.yesaladin.shop.product.domain.repository.SearchProductRepository;
 import shop.yesaladin.shop.product.dto.SearchedProductResponseDto;
 
@@ -36,12 +36,12 @@ class SearchProductServiceImplTest {
                 .title("title")
                 .discountRate(10)
                 .sellingPrice(1000L)
-                .authors(List.of(new SearchedAuthor(1L, "author")))
+                .authors(List.of(new SearchedProductAuthor(1L, "author")))
                 .isForcedOutOfStack(false)
-                .thumbnailFileUrl(new SearchedFile(1L, "깃 허브.jpg", LocalDate.now()))
+                .thumbnailFileUrl(new SearchedProductFile(1L, "깃 허브.jpg", LocalDate.now()))
                 .publishedDate(LocalDate.now().toString())
-                .categories(List.of(new SearchedCategories(12L, null, "국내소설", true, false)))
-                .tags(List.of(new SearchedTags(1L, "tag1")))
+                .categories(List.of(new SearchedProductCategory(12L, null, "국내소설", true, false)))
+                .tags(List.of(new SearchedProductTag(1L, "tag1")))
                 .build();
 
         dummy = List.of(dto);
